pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-segmentation-scratch"
version = "0.1.0"
description = "Binary semantic segmentation from scratch with PyTorch."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.7",
  "numpy>=1.24",
  "Pillow>=10.0",
  "PyYAML>=6.0",
  "torch",
  "torchvision",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
